refactor compiler detect (#7777)
* refactor compiler detect

* fix output string
